google-logo
removed apple sign up and logo and in both signup page and login page and added google signup in login page
